effects modularization
